requirements.txt
torch
numpy
opencv-python
Pillow
tqdm
submodlib
git+https://github.com/openai/CLIP.git
